Rename check-username.ts
